refactor: move route logic into controllers and services
