+++ apps/server/src/controllers/health.js
@@ -1,0 +1,5 @@
+import * as healthService from '../services/health.js';
+
+export function checkHealth(_req, res) {
+  res.json(healthService.check());
+}
