+++ apps/server/src/routes/voice.js
@@ -1,24 +1,7 @@
 import { Router } from 'express';
-import Twilio from 'twilio';
-import { serverEnv as env } from '@shared/env';
 import { verifyTwilioSignature } from '../middleware/verifyTwilio.js';
+import * as voiceController from '../controllers/voice.js';
 
 export const voice = Router();
 
-voice.post('/voice/outbound', verifyTwilioSignature, (req, res) => {
-  const twiml = new Twilio.twiml.VoiceResponse();
-  const to = (req.body.To || '').trim();
-
-  if (!to) {
-    twiml.say('Missing destination.');
-    return res.type('text/xml').status(200).send(twiml.toString());
-  }
-
-  const dial = twiml.dial({ callerId: env.callerId });
-  if (to.startsWith('client:')) {
-    dial.client({}, to.replace('client:', ''));
-  } else {
-    dial.number({}, to);
-  }
-  return res.type('text/xml').status(200).send(twiml.toString());
-});
+voice.post('/voice/outbound', verifyTwilioSignature, voiceController.outbound);
