+++ apps/server/src/services/health.js
@@ -1,0 +1,1 @@
+export const check = () => ({ ok: true });
